+++ agent/agent.go
@@ -1,6 +1,7 @@
 package agent
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"io"
@@ -25,7 +26,7 @@ const (
 // dialer is the websocket dialer used to connect to a gateway server.
 var dialer = websocket.DefaultDialer
 
-// Agent remote-listens on a port on a gateway server.
+// Agent tunnels remote port on a gateway server to local destination.
 type Agent struct {
 	gatewayURL  string
 	key         string
@@ -33,8 +34,9 @@ type Agent struct {
 	destination string
 }
 
-// Start starts agents with given configuration.
-func Start(conf config.Agent) error {
+// Start starts tunneling agents with given configurations. The agents and
+// tunneled connections are cancelable via context.
+func Start(conf config.Agent, ctx context.Context) error {
 	if err := conf.Check(); err != nil {
 		return fmt.Errorf("config error: %w", err)
 	}
@@ -50,12 +52,15 @@ func Start(conf config.Agent) error {
 		}
 
 		tasks.Go(func() error {
-			delay := time.Tick(retryInterval)
+			retry := time.NewTicker(retryInterval)
+			defer retry.Stop()
 			for {
-				if err := agent.Start(); err != nil {
-					log.Printf("Agent error %q - recovering...", err)
-					<-delay
+				err := agent.Start(ctx)
+				if err == nil || !isRecoverable(err) {
+					return err
 				}
+				log.Printf("Agent error %q - recovering...", err)
+				<-retry.C
 			}
 		})
 	}
@@ -63,18 +68,53 @@ func Start(conf config.Agent) error {
 	return tasks.Wait()
 }
 
-// Start starts remote-listening.
-func (agent Agent) Start() error {
-	url := agent.gatewayURL + "/" + agent.service.Protocol + "/" + strconv.Itoa(agent.service.Port)
+// Returns true if agent error is recoverable (by restarting agent).
+func isRecoverable(err error) bool {
+	if errors.Is(err, context.Canceled) {
+		return false
+	}
+	return true
+}
 
+// hookCancel launches a goroutine for handling task cancellation. handler is
+// called upon cancellation. The caller must invoke returned function after the
+// cancelable task is finished.
+func hookCancel(ctx context.Context, handler func()) func() {
+	end := make(chan struct{})
+	unhook := func() {
+		close(end)
+	}
+	go func() {
+		select {
+		case <-ctx.Done():
+			handler()
+		case <-end:
+		}
+	}()
+	return unhook
+}
+
+// Start connects to tunneling server and starts listening on the remote ports.
+// The connection to the server and tunnels can be canceled via context.
+func (agent *Agent) Start(ctx context.Context) error {
+
+	// Connection on the remote port is first notified to us as a WebSocket
+	// message. We then call back server to create another WebSocket channel
+	// for tunneling the connection (in Agent.tunnel function).
+
+	url := agent.gatewayURL + "/" + agent.service.Protocol + "/" + strconv.Itoa(agent.service.Port)
 	header := http.Header{}
 	header.Add("Authorization", "Bearer "+agent.key)
-
-	ws, _, err := dialer.Dial(url, header)
+	ws, _, err := dialer.DialContext(ctx, url, header)
 	if err != nil {
 		return err
 	}
 	defer closeWebsocket(ws)
+
+	unhookCancel := hookCancel(ctx, func() {
+		closeWebsocket(ws)
+	})
+	defer unhookCancel()
 
 	log.Printf("Listening on remote port: %s", agent.service)
 
@@ -85,40 +125,61 @@ func (agent Agent) Start() error {
 				log.Printf("Agent %s closed normally.", agent.service)
 				return nil
 			}
+			if errors.Is(err, net.ErrClosed) {
+				log.Printf("Agent %s is canceled.", agent.service)
+				return context.Canceled
+			}
 			return err
 		}
 
 		go func() {
-			if err := agent.tunnel(accept); err != nil {
+			if err := agent.tunnel(accept, ctx); err != nil {
 				log.Printf("Tunneling error: %s", err)
 			}
 		}()
 	}
 }
 
-func (agent Agent) tunnel(accept service.BinderAcceptMessage) error {
+// tunnel proxies accepted connection on the remote tunneling server to a local
+// connection to the forwarding destination. The tunnel can be canceled via
+// context.
+func (agent *Agent) tunnel(accept service.BinderAcceptMessage, ctx context.Context) error {
 	log.Printf(
 		"Tunneling remote connection from %s to %s",
 		accept.PeerAddress,
 		agent.destination,
 	)
 
+	// Local connection to the forwarding destination.
 	conn, err := net.Dial(agent.service.Protocol, agent.destination)
 	if err != nil {
 		return err
 	}
 	defer conn.Close()
 
+	// Remote connection proxied through WebSocket.
 	url := agent.gatewayURL + "/session/" + accept.SessionID
-	ws, _, err := dialer.Dial(url, nil)
+	ws, _, err := dialer.DialContext(ctx, url, nil)
 	if err != nil {
 		return err
 	}
 	defer closeWebsocket(ws)
 
+	unhookCancel := hookCancel(ctx, func() {
+		conn.Close()
+		closeWebsocket(ws)
+	})
+	defer unhookCancel()
+
 	tasks := taskch.New()
 
-	// Uplink
+	// The tunnel looks like this:
+	//
+	// Client <------> Server <------> Agent <------> Destination
+	//                           ws            conn
+	//
+
+	// Uplink: (Client --->) Server ---> Agent ---> Destination
 	tasks.Go(func() error {
 		buf := make([]byte, config.BufferSize)
 
@@ -134,7 +195,7 @@ func (agent Agent) tunnel(accept service.BinderAcceptMessage) error {
 		}
 	})
 
-	// Downlink
+	// Downlink: (Client <---) Server <--- Agent <--- Destination
 	tasks.Go(func() error {
 		buf := make([]byte, config.BufferSize)
 
@@ -155,6 +216,15 @@ func (agent Agent) tunnel(accept service.BinderAcceptMessage) error {
 	if errors.Is(err, io.EOF) {
 		log.Printf(
 			"Destination closed. Finishing session %s -> %s",
+			accept.PeerAddress,
+			agent.destination,
+		)
+		return nil
+	}
+
+	if errors.Is(err, net.ErrClosed) {
+		log.Printf(
+			"Canceled. Finishing session %s -> %s",
 			accept.PeerAddress,
 			agent.destination,
 		)